default stops creator
